pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pipeworks-image-generator"
version = "0.1.0"
description = "Programmatic image generation with Z-Image-Turbo and extensible plugin system"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "aapark"}
]
dependencies = [
    "torch>=2.5.0",
    "diffusers>=0.32.0",
    "gradio>=5.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pillow>=10.0.0",
    "accelerate>=0.30.0",
    "transformers>=4.40.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[project.scripts]
pipeworks = "pipeworks.ui.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
